Add function Paytm_GetFreeROM to get free flash size of device

+++ Code/sdk/app/paytm_api/inc/paytm_dev_api.h
@@ -14,12 +14,7 @@
 
 int64 Paytm_GetPhysicalROM(void);
 int64 Paytm_GetTotalROM(void);
-
-/**
- * @description: Return free rom size(internel flash) of device, uints is byte
- * @return {*}
- */
-int32 Paytm_GetFreeROM(void);
+int64 Paytm_GetFreeROM(void);
 uint16 Paytm_GetTotalNVRAM(void);
 uint16 Paytm_GetFreeNVRAM(void);
 int64 Paytm_GetTotalHeapSize(void);
